refactor sparse_sim.rs to sparse_stab_bindings.rs
