+++ crates/pecos-python/src/sparse_stab_bindings.rs
